Add files via upload
The Pokedex program was designed as if it it were an actual Pokedex to be used for actual Pokemon battles. Say, for example, you wanted to try for the Pewter City gym badge by defeating its gym leader, Brock. You know that Brock's first Pokemon is a level 12 Geodude. You have a level 10 Butterfree and a level 8 Nidoran. Which of your Pokemon has a better chance of beating Brock's Geodude? That's where the Pokedex comes in.

It works like this. You first pick the Pokemon you will use in the fight, let's say Butterfree. Then you select the Pokemon you will be battling: Geodude. You declare the level of each Pokemon--10 and 12, respectively. Next you list the set of moves each Pokemon has. In our hypothetical example we are assuming you evolved the Butterfree from Caterpie so the move list is Tackle, String Shot, Harden, and Confusion. Now you list the moves the opponent knows: Tackle and Defense Curl for Geodude. The Pokedex takes this data and has the two Pokemon battle it out 1000 times. In the end, the Pokedex will display your chance of winning the battle. For this scenario, we have a 76.2% of defeating Brock's Geodude with our Butterfree.

Now we can see if we have a better chance of using Nidoran instead of Butterfree. The moves available for a level 8 Nidoran are Tackle, Leer, and Horn Attack. The move list will be the same for the opponent because Brock is still using his level 12 Geodude. After passing through the data, the Pokedex concludes that we have a 0% chance of defeating Geodude with our Nidoran. Yikes! Best to stick with Butterfree, then.

So how does the Pokedex actually work? First, it will ask which Pokemon you want to use. You'll need to know what the Pokedex number of your Pokemon is in order for the program to select the right Pokemon. In our example, if we choose Butterfree, we would input the number 12, which is Butterfree's Pokedex number. This is because when I stored the information of each Pokemon in a MySQL database, I used the Pokedex number as the primary key. Then the Pokedex asks which Pokemon the opponent is using. Since Brock uses Geodude, we enter 74, which is Geodude's Pokedex number.

The next thing the program asks for is the level of the Pokemon. While I didn't put any restrictions on the number you can choose (meaning the Pokedex will allow you have a level 10,000 Butterfree), Pokemon the game only allows levels ranging from 2-100. If you really wanted to cheat here you can, it would just make your battle data useless. After picking your level, you input the level of your opponent's Pokemon.

Finally, the Pokedex prompts you for the moves your Pokemon knows and then asks for the four moves your opponent knows. The way I programmed the Pokedex, it assumes a move set of four moves, and caps it at four, meaning you can't have a Butterfree with seventeen moves. (The game Pokemon uses this four-move restriction and I copied that philosophy here.) But because there is an assumption of four moves, in the cases where your Pokemon only knows three moves depending on its level, you would have to enter one of the moves twice. For example, our level 8 Nidoran only knows three moves: Tackle, Leer, and Horn Attack. When I entered the list of moves, I input Tackle, Leer, Horn Attack, and Horn Attack. (I favored Horn Attack because it was the strongest attack that a level 8 Nidoran would know.)

After inputting the information, the Pokedex has the two Pokemon fight each other in a loop for 1000 matches. The attacks are chosen at random. When the battle loop ends, your wins are displayed as a percentage out of 100. 

I only created a database of the original 151 Pokemon and included every move to choose from except for Flail. Flail isn't a move in itself per se, but rather a side effect of when a learned move runs out of PP. What is PP? In the Pokemon game, PP is how many times you can use a move in a battle. The move Tackle has a PP of 40. The move Fire Blast has a PP of 5. You can use Tackle 40 times in a single battle and can only use Fire Blast 5 times. I did not program any PP restrictions into the Pokedex. I wanted it to feel like it would be an actual Pokemon battle in real life. If Pokemon were real creatures, there would still be the move restrictions (a Fire-type Pokemon would not be able to use a Water-type move, for example, and there would still be the four-move cap), but there would be no restrictions on how many times a Pokemon could use a move that it knows. I followed that philosophy in creating the Pokedex.

I hope you have fun playing around with the Pokedex and using it to give you an advantage in your battles. Best of luck, trainers!